Move commands and CLI entry points to `spdl/autoresearch/pipeline_optimization/` (#1453)
Summary:
Pull Request resolved: https://github.com/facebookresearch/spdl/pull/1453

Move `utils/commands/` → `_impl/_commands/` (5 command modules), and CLI entry points `cli.py` → `_impl/_cli.py`, `cmd.py` → `_impl/_cmd.py`, `run.py` → `_impl/_run.py`.

`launch.sh` and `README.md` stay in `spdl/tools/autoresearch/` — they are entry point artifacts, not Python library code.

Update `spdl/tools/autoresearch/BUCK` binary targets to use new `main_module` paths. Update all test and fb/ BUCK files to depend on `//spdl/autoresearch:public_api` and the private target.

Differential Revision: D105981372
