+++ src/spdl/autoresearch/pipeline_optimization/__init__.py
@@ -4,6 +4,13 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-"""SPDL pipeline optimization workflow for autoresearch."""
+"""SPDL pipeline optimization implementation of autoresearch.
 
-__all__: list[str] = []
+Provides the complete workflow for automatically optimizing SPDL data loading
+pipelines: supervisor CLI, engine runner, experiment adapter, platform
+providers, and CLI commands.
+"""
+
+from ._cli import main
+
+__all__ = ["main"]
