+++ src/spdl/autoresearch/pipeline_optimization/_commands/_report.py
