+++ src/spdl/autoresearch/pipeline_optimization/_commands/_status.py
