+++ tools/autoresearch/tests/test_autoresearch_workflow.py
@@ -25,6 +25,8 @@ from spdl.autoresearch.core import (
     HypothesisNode,
     TaskSpec,
 )
+from spdl.autoresearch.pipeline_optimization._commands._report import _read_failures
+from spdl.autoresearch.pipeline_optimization._commands._status import _failure_summary
 from spdl.autoresearch.pipeline_optimization._ops import (
     _WorkflowStateStore,
     PipelineOptimizationWorkflow,
@@ -60,8 +62,6 @@ from spdl.autoresearch.pipeline_optimization._platform import (
 )
 from spdl.autoresearch.pipeline_optimization._platform._agents import _MockAgent
 from spdl.autoresearch.pipeline_optimization._platform._local import _summarize_error
-from spdl.tools.autoresearch.utils.commands.report import _read_failures
-from spdl.tools.autoresearch.utils.commands.status import _failure_summary
 
 __all__: list[str] = []
