+++ tools/autoresearch/tests/test_platform.py
@@ -12,6 +12,7 @@ import unittest
 from pathlib import Path
 from unittest.mock import patch
 
+from spdl.autoresearch.pipeline_optimization._commands import _queue as cmd_queue
 from spdl.autoresearch.pipeline_optimization._platform import (
     _MetricsEvidence,
     AutoresearchPlatform,
@@ -21,7 +22,6 @@ from spdl.autoresearch.pipeline_optimization._platform._agents import (
     _MockAgent,
     _parse_agent_result,
 )
-from spdl.tools.autoresearch.utils.commands import queue as cmd_queue
 
 __all__: list[str] = []
