--- tools/autoresearch/utils/__init__.py
@@ -1,15 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the BSD-style license found in the
-# LICENSE file in the root directory of this source tree.
-
-"""Autoresearch utility package.
-
-Implementation code is organized by boundary:
-``workflow`` for autoresearch orchestration, ``platform`` for platform
-integration, ``infra`` for low-level helpers, and ``commands`` for CLI
-subcommands.
-"""
-
-__all__: list[str] = []
